Refactor JobType as interface and AbstractJobType
