+++ kernel/data-pipeline/api/src/main/java/org/apache/shardingsphere/data/pipeline/api/job/AbstractJobType.java
@@ -19,43 +19,27 @@ package org.apache.shardingsphere.data.pipeline.api.job;
 
 import com.google.common.base.Preconditions;
 import lombok.Getter;
-
-import java.util.Arrays;
-import java.util.Map;
-import java.util.stream.Collectors;
+import org.apache.shardingsphere.data.pipeline.spi.job.JobType;
 
 /**
- * Job type.
+ * Abstract job type.
  */
 @Getter
-public enum JobType {
+public abstract class AbstractJobType implements JobType {
     
+    // TODO delete
+/*
     MIGRATION("MIGRATION", "01"),
     CONSISTENCY_CHECK("CONSISTENCY_CHECK", "02");
-    
-    private static final Map<String, JobType> CODE_JOB_TYPE_MAP;
-    
-    static {
-        CODE_JOB_TYPE_MAP = Arrays.stream(JobType.values()).collect(Collectors.toMap(JobType::getTypeCode, each -> each));
-    }
+*/
     
     private final String typeName;
     
     private final String typeCode;
     
-    JobType(final String typeName, final String typeCode) {
+    AbstractJobType(final String typeName, final String typeCode) {
         this.typeName = typeName;
         Preconditions.checkArgument(2 == typeCode.length(), "code length is not 2");
         this.typeCode = typeCode;
-    }
-    
-    /**
-     * Value of by code.
-     *
-     * @param typeCode type code
-     * @return job type, might be null
-     */
-    public static JobType valueOfByCode(final String typeCode) {
-        return CODE_JOB_TYPE_MAP.get(typeCode);
     }
 }
